+++ inc/grMadgwick.h
@@ -17,20 +17,20 @@ class GRMadgwick
 		GRMadgwick& operator=(const GRMadgwick&);
 
 	
-    	void MadgwickAHRSupdate(float gx, float gy, float gz, float ax, float ay, float az, 
-                float mx, float my, float mz, std::vector<float>*);
-        void MadgwickAHRSupdateIMU(float gx, float gy, float gz, float ax, float ay, float az, 
-                std::vector<float>*);
+    	void MadgwickAHRSupdate(double gx, double gy, double gz, double ax, double ay, double az, 
+                double mx, double my, double mz, std::vector<double>*);
+        void MadgwickAHRSupdateIMU(double gx, double gy, double gz, double ax, double ay, double az, 
+                std::vector<double>*);
         bool setFreqCalibration(int);	
 	private:
 		double constrain(double , double , double );
         double gravity = 256;
         //madgwick algorithm vars and methods
-        float q0 , q1, q2, q3;  // quaternion of sensor frame relative to auxiliary frame
-        float beta;
+        double q0 , q1, q2, q3;  // quaternion of sensor frame relative to auxiliary frame
+        double beta;
         int freqCallibration;
 
-        std::vector<float> computeAngles();
-        float invSqrt(float x);
+        std::vector<double> computeAngles();
+        double invSqrt(double x);
 };
 #endif
